Move InputManager to new input module/file
