+++ app/src/main.rs
@@ -22,6 +22,9 @@ use rust_wii_lib::{Changes, Controls, GameState};
 pub mod renderer;
 use renderer::*;
 
+pub mod input;
+use input::InputManager;
+
 mod raw_data_store;
 
 mod target_tests;
@@ -68,8 +71,7 @@ fn main_game() -> isize {
 
         renderer.render_world(&game_state.world);
     }
-    shutdown();
-    0
+    shutdown()
 }
 
 /// Registers the power callback,
@@ -95,34 +97,5 @@ pub fn shutdown() -> ! {
     unsafe {
         STM_ShutdownToStandby();
     }
-}
-
-struct InputManager {
-    wii_mote: Input,
-}
-
-impl InputManager {
-    pub fn new() -> Self {
-        // Setup the wiimote
-        Input::init(ControllerType::Wii);
-        let wii_mote = Input::new(ControllerType::Wii, ControllerPort::One);
-        wii_mote
-            .as_wpad()
-            .set_data_format(WPadDataFormat::ButtonsAccelIR);
-        Self { wii_mote }
-    }
-
-    pub fn update(&mut self) -> Controls {
-        Input::update(ControllerType::Wii);
-        Controls {
-            home_button_down: self.wii_mote.is_button_down(Button::Home),
-            one_button_down: self.wii_mote.is_button_down(Button::One),
-        }
-    }
-}
-
-impl Default for InputManager {
-    fn default() -> Self {
-        Self::new()
-    }
+    core::unreachable!()
 }
